# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SimpliGPS API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/devices": {
            "get": {
                "summary": "Display a listing of the Devices.",
                "operationId": "App\\Http\\Controllers\\DevicesController::index",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Display a paginated response of devices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/device"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    },
                    {
                        "tenant": []
                    }
                ]
            }
        },
        "/api/v1/me/accounts": {
            "get": {
                "summary": "Devuelve todas las cuentas de un usuario, necesario para el login y la especificacion del tenant.",
                "operationId": "App\\Http\\Controllers\\MeController::accounts",
                "responses": {
                    "200": {
                        "description": "Devuelve todas las cuentas de un usuario, necesario para el login y la especificacion del tenant.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "device": {
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array."
                    },
                    "name": {
                        "description": "Transform the resource into an array."
                    },
                    "gps": {
                        "description": "Transform the resource into an array."
                    },
                    "brand": {
                        "description": "Transform the resource into an array."
                    },
                    "model": {
                        "description": "Transform the resource into an array."
                    },
                    "internal_number": {
                        "description": "Transform the resource into an array."
                    },
                    "truck": {
                        "description": "Transform the resource into an array."
                    },
                    "position": {
                        "description": "Transform the resource into an array."
                    },
                    "reference_data": {
                        "description": "Transform the resource into an array."
                    },
                    "is_connected": {
                        "description": "Transform the resource into an array."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://api.simpligps.test/oauth/authorize",
                        "tokenUrl": "http://api.simpligps.test/oauth/token",
                        "refreshUrl": "http://api.simpligps.test/token/refresh",
                        "scopes": []
                    }
                }
            },
            "tenant": {
                "type": "apiKey",
                "name": "X-Tenant-Id",
                "description": "Each account is a tenant and has an UUID that needs to be attached in each request",
                "in": "header"
            }
        }
    }
}